Data file changed to avoid error messages.
There was an annoying error when using gnss0.bin. This was caused by changed number of satellites because of measurement validity checks. Now the problem is just bypassed by using another data file, i.e. gnsa14.bin, which was collected in London. As a side note, if you could successfully use gnss0.bin, you could find that it was collected in Shanghai.